[refactor] 그룹 및 초대 서비스 리팩토링
- GroupService 및 InvitationService 인터페이스로 변경하여 구현체와 분리
- GroupServiceImpl 및 InvitationServiceImpl 클래스 추가로 비즈니스 로직 구현
- GroupHandler 및 MemberHandler를 통한 예외 처리 통합
- InvitationController 및 InvitationService의 초대 생성 로직 개선
- MemberDataInitializer에서 더미 사용자 및 그룹 데이터 초기화 로직 추가
