+++ src/main/java/com/mymate/mymate/group/dto/InvitationCreateRequest.java
@@ -1,17 +1,20 @@
 package com.mymate.mymate.group.dto;
 
-import jakarta.validation.constraints.NotNull;
+import io.swagger.v3.oas.annotations.media.Schema;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
+@Schema(description = "그룹 초대 생성 요청")
 public class InvitationCreateRequest {
 
-    @NotNull(message = "초대받을 사용자 ID는 필수입니다")
-    private Long inviteeId;
+    @NotBlank(message = "초대받을 사용자 식별자는 필수입니다")
+    @Schema(description = "초대받을 사용자 식별자 (사용자 ID 또는 이메일)", example = "SZZYDE770", required = true)
+    private String inviteeIdentifier;  // 사용자 ID 또는 이메일 (예: "SZZYDE770" 또는 "user@gmail.com")
 
-    public InvitationCreateRequest(Long inviteeId) {
-        this.inviteeId = inviteeId;
+    public InvitationCreateRequest(String inviteeIdentifier) {
+        this.inviteeIdentifier = inviteeIdentifier;
     }
 }
