+++ src/main/java/com/mymate/mymate/member/dto/MemberSearchResponse.java
@@ -1,22 +1,21 @@
 package com.mymate.mymate.member.dto;
 
 import com.mymate.mymate.member.Member;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
+@AllArgsConstructor
 public class MemberSearchResponse {
-
     private Long id;
+    private String userId;
     private String username;
-    private String name;
     private String email;
-
+    
     public MemberSearchResponse(Member member) {
         this.id = member.getId();
+        this.userId = member.getUserId();
         this.username = member.getUsername();
-        this.name = member.getUsername();
         this.email = member.getEmail();
     }
 }
