+++ src/main/resources/application.yml
@@ -17,6 +17,11 @@ spring:
     redis:
       host: ${REDIS_HOST:localhost}
       port: ${REDIS_PORT:6379}
+  h2:
+    console:
+      enabled: true
+      settings:
+        web-allow-others: true
 
 # 로그 설정
 logging:
